requirements.txt
biopython==1.68
argparse==1.1
behave==1.2.5
scripttest==1.3
